Feat: v3 Refactor (#361)
* Huge refactor

* Sessions and transactions 🎉

* Fixes from code review

* Simplify integration types

* A few fixes and improvements

* Fix the `event_type` tags

* Make main process IPC a non-optional feature

* A few fixes

* Fix linting

* update yarn.lock

* Native renderer crashes now call before send

* Track native crashes as crashed session

* Remove crashes directory between tests

* re-enable retries 🤷‍♂️

* changes from PR review

* Fix issue with missing remote modules in Electron v15 types

* Final changes

+++ test/e2e/process.ts
@@ -19,10 +19,14 @@ export class ProcessStatus {
   /** Determines whether this process is still running. */
   public async isRunning(): Promise<boolean> {
     try {
-      process.kill(this._chProcess.pid, 0);
-      return true;
+      if (this._chProcess.pid) {
+        process.kill(this._chProcess.pid, 0);
+        return true;
+      }
     } catch (e) {
-      return false;
+      //
     }
+
+    return false;
   }
 }
